Fix sorting and searching
Sorting broke because pagination was skipped.  Searching was unstable
because listings with different categories but same tags were shown
together.  This update causes search to be performed only for tags,
not titles or item descriptions.  If no matching tags are found, then
all listings are returned.  However, if listing titles were stored as
tags, then the search would also return listings with matching titles
for free.